+++ src/main/java/dev/abarmin/aml/dashboard/block/avatar/AvatarBlockProps.java
@@ -3,6 +3,7 @@ package dev.abarmin.aml.dashboard.block.avatar;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import dev.abarmin.aml.dashboard.domain.BlockProps;
 import dev.abarmin.aml.file.FileId;
+import dev.abarmin.aml.file.FileService;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -20,6 +21,18 @@ public class AvatarBlockProps implements BlockProps {
 
   private FileId avatarId;
   private FileId backgroundId;
+
+  @JsonIgnore
+  public Map<String, String> getCoverStyle(FileService fileService) {
+    final String backgroundPublicUrl = fileService.getPublicUrl(backgroundId);
+
+    return Map.of(
+      "background-image", "url(" + backgroundPublicUrl + ")",
+      "background-position", "center",
+      "background-repeat", "no-repeat",
+      "height", "200px"
+    );
+  }
 
   @JsonIgnore
   public Map<String, String> getAvatarStyle() {