+++ src/main/java/dev/abarmin/aml/dashboard/domain/Page.java
@@ -13,7 +13,21 @@ public record Page(
   @Column("page_title") String title,
   @Column("page_home") boolean isHome,
   @Column("page_deleted") boolean isDeleted,
+  @Column("page_props") PageProps pageProps,
   @Column("created_at")Instant createdAt,
   @Column("updated_at")Instant updatedAt
   ) {
+
+  public Page withProps(PageProps props) {
+    return new Page(
+      id,
+      profileId,
+      title,
+      isHome,
+      isDeleted,
+      props,
+      createdAt,
+      Instant.now()
+    );
+  }
 }