+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed padding between fields in the header block props. 
 * Pages now have background color. 
+* `l`-page has its own style now. 
 
 # 0.0.16