+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.17
+
+* Fixed padding between fields in the header block props. 
+* Pages now have background color. 
+* `l`-page has its own style now. 
+* Added a banner block to the end of the page. 
+
 # 0.0.16
 
 * Split components and their props into separate files.